refactor: 카드컨테이너 수정
